pyproject.toml
[tool.poetry]
name = "langchain-playground"
version = "0.1.0"
description = ""
authors = ["Jack Michaud <jack@lomz.me>"]
readme = "README.md"
packages = [{include = "langchain_playground"}]

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.108"
python-dotenv = "^1.0.0"
openai = "^0.27.0"
google-search-results = "^2.4.1"
unstructured = {extras = ["local-inference"], version = "^0.5.1"}
pypdf = "^3.5.0"
pinecone-client = "^2.2.1"
tiktoken = "^0.3.0"
chromadb = "^0.3.11"

[tool.poetry.extras]
load-raw-data = ["unstructured", "pypdf"]
google-search = ["google-search-results"]


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
